Completely remove Kinds module
