+++ frontend/src/Language/Granule/Checker/Patterns.hs
@@ -12,7 +12,6 @@ import Language.Granule.Checker.Types (equalTypesRelatedCoeffectsAndUnify, SpecI
 import Language.Granule.Checker.Flatten
 import Language.Granule.Checker.Monad
 import Language.Granule.Checker.Predicates
-import Language.Granule.Checker.Kinds
 import Language.Granule.Checker.SubstitutionContexts
 import Language.Granule.Checker.Substitution
 import Language.Granule.Checker.Variables
